Split codegen.rs into items, types, and dts sub-modules
Reduces codegen.rs from 1856 to 751 lines by extracting:

- codegen/items.rs (705 lines): item emission (import, const, fn,
  for block, test block, type decl, derived display)
- codegen/dts.rs (317 lines): .d.ts declaration stub generation
- codegen/types.rs (99 lines): type expression emission

Joins existing codegen/expr.rs, codegen/jsx.rs, codegen/match_emit.rs.
Purely structural - no behavioral changes.
